requirements.txt
pandas
datasets
transformers
bitsandbytes
accelerate
python-dotenv

